==> Data Analytics and Linear Regression/1. Churn Dataset - Variable Identification.ipynb <==
# Churn Dataset Analysis

### Basic Information
- **Dataset Shape**: The dataset contains 15,929 rows and 19 columns.
- **Data Types**: The columns comprise a mix of numerical (`float64`) and categorical data (`object` for strings).
- **Missing Values**: There are missing values in the dataset.
- **Unique Values**: The dataset has a wide range of unique values across different columns, indicating a diverse dataset.

### Columns Overview
- **`vintage`**: Customer relationship duration with the bank (1,332 unique values).
- **`age`**: Age of the customer (90 unique values).
- **`gender`**: Gender of the customer (2 unique values - presumably male and female).
- **`dependents`**: Number of dependents of the customer (14 unique values).
- **`occupation`**: Occupation of the customer (5 unique categories).
- **`city`**: City codes (1,288 unique values).
- **`customer_nw_category`**: Customer net worth category (3 unique categories).
- **`branch_code`**: Code of the branch (2,806 unique values).
- **Financial columns**: Several columns like `current_balance`, `previous_month_end_balance`, `average_monthly_balance_prevQ`, and similar others indicating various aspects of the customer's account balance and transactions.
- **`churn`**: Customer churn status (2 unique values, likely indicating churned or not churned).

## Reading Files into Python

# importing libraries
import pandas as pd

pwd  

#importing data
file_path = r"C:\Users\Prashant\Downloads\AV\bbplus"  # folder location of your dataset
data = pd.read_csv(file_path + r'\churn_prediction.csv')

data = pd.read_csv('churn_prediction.csv')  
# this code works only when the CSV file is present in the same folder as THIS 
# Jupyter notebook

#first 5 instances using "head()" function
data.head()

#last 10 instances using "tail()" function
data.tail(10)

#finding out the shape of the data using "shape" variable: Output (rows, columns)
data.shape

#Printing all the columns present in data
data.columns

## Variable Identification and Typecasting

data.info()

# A closer look at the data types present in the data
data.dtypes

There are a lot of variables visible at one, so let's narrow this down by looking **at one datatype at once**. We will start with int


### Integer Data Type

# Identifying variables with integer datatype
sum(data.dtypes == "int64" )  # we have 5 columns with int64 datatype

data.dtypes[data.dtypes == "int64"]

integer_columns = data.select_dtypes(include=['int64', 'float64'])
integer_columns  # part of the original dataframe including ONLY the "int64" & "float64" columns 

Summary:

*    **Customer id** are a unique number assigned to customers. It is **Okay to consider this as an Integer**. This variable would not considered for our analysis.

*    **branch code** again represents different branches, therefore it should be **convereted to category**.

*    **Age** and **Vintage** are also numbers and hence we are okay with them as integers.

*    **customer_networth_category** is supposed to be an ordinal category, **should be converted to category**.

*    **churn** : 1 represents the churn and 0 represents not churn. However, there is no comparison between these two categories. This **needs to be converted to category datatype**.


data.describe()

data["churn"].dtype

# converting churn, branch code and customer_nw_category  to "category"
data['churn'] = data['churn'].astype('category')
data['branch_code'] = data['branch_code'].astype('category')
data['customer_nw_category'] = data['customer_nw_category'].astype('category')

data.dtypes

data.dtypes[data.dtypes == 'int64']

data.info()

### Float Data Type

# Identifying variables with float datatype
data.dtypes[data.dtypes == 'float64']

Summary:

*    **dependents** is expected to be a whole number. **Should be changed to integer type**

*    **city** variable is also a unique code of a city represented by some interger number. **Should be converted to Category type**

*    Rest of the variables like **credit, balance and debit** are best represented by the float variables.

# converting "dependents" and "city" to their respective types
data['dependents'] = data['dependents'].astype("Int64")
data['city'] = data['city'].astype('category')

# checking
data[['dependents','city']].dtypes

### Object Data Type

data.dtypes

*    **variables like 'gender', 'occupation' and 'last_transaction' are of type object**. This means that **Pandas was not able to recognise the datatype** of these three variables.

# Manually checking object types
data[['gender','occupation','last_transaction']].head(7)

*    **gender** and **occupation** variables **belong to categorical data types**.
*    **last_transaction** should be a  **datetime variable**.

# typecasting "gender" and "occupation" to category type
data['gender'] = data['gender'].astype('category')
data['occupation'] = data['occupation'].astype('category')

# checking
data[['gender','occupation']].dtypes

data.info()

100*(2.6-2.1)/2.6 

### datetime Data Type

# Convert the "last_transaction" column into datetime object.
data['last_transaction'] = pd.to_datetime(data['last_transaction'])

data.dtypes

data.info()

# extracting new columns from "last_transaction"  >> Feature Engineering

# last day of year when transaction was done
data['doy_ls_tran'] = data.last_transaction.dt.dayofyear

# week of year when last transaction was done
data['woy_ls_tran'] = data.last_transaction.dt.isocalendar().week

# month of year when last transaction was done
data['moy_ls_tran'] = data.last_transaction.dt.month

# day of week when last transaction was done
data['dow_ls_tran'] = data.last_transaction.dt.dayofweek

# checking new extracted columns using datetime
data[['last_transaction','doy_ls_tran','woy_ls_tran','moy_ls_tran','dow_ls_tran']].head()

The first column is the complete date of the last transaction which was done by the any given customer.

The next columns represent the day of year, week of year, month of year, day of week when the last transaction was done.

**Breaking down the date variable** into these granular information will **help us in understand when the last transaction was done from different perspectives**. Now that we have extracted the essentials from the last_transaction variables, we will drop it from the dataset.



# Drop the last_transaction column
# data = data.drop(columns = ['last_transaction'])
# data.drop(data['last_transaction', axis=1, inplace=True)

data.drop('last_transaction', axis=1)

data.columns

data.drop('last_transaction', axis=1, inplace=True)

data.dtypes

data.info()

# Save your dataset for further analysis
data.to_csv("customer_churn_v1.csv", index=False)
data.to_pickle("customer_churn_v1.pkl")
