==> Data Analytics and Linear Regression/2. Churn Dataset - preprocessing.ipynb <==
# Continue Data Analysis

import pandas as pd

# Read from the CSV file
df = pd.read_csv("customer_churn_v1.csv")

df.info()

df.head()

# Read from the pickle file v1
df = pd.read_pickle("customer_churn_v1.pkl")

df.head()

df.info()

pd.set_option('display.precision', 2)

df.describe().T # quick statistical summary of the numeric variables

# statistical summary for the categorical columns
df.describe(include='category')

100*288/9308  # 3% missing vlaues

100*47/9814  # less than 0.5% missing vlaues

100*460/1949  # 24 % of the top most value

9308/len(df)  # top most vlaue in gender is already 58% of the dataset

9814/len(df)  # top most vlaue in occupation is already 61% of the dataset

100*1949/len(df)  # top most vlaue in city is ONLY 12% of the dataset

# Missing values

# list of all the columns having missing vlaues & the number f missing values
df.isna().sum()

# get the number of missing values for gender column only
df['gender'].isnull().sum()

df['gender']

df['gender'].mode()[0]

# TO DO: Fill the missing values of gender column with the mode value
df['gender'].fillna(value = df['gender'].mode()[0], inplace=True)

df['gender'].isna().sum()   # there no more missing vlues

df['occupation'].mode()[0]

# TO DO: Fill the missing values of occupation column with the mode value
df['occupation'].fillna(value = df['occupation'].mode()[0], inplace=True)

df['occupation'].isna().sum() 

df.city.nunique()

## Missing values in the depedents Col

df['dependents'].mean()

df['dependents'].value_counts(normalize=True)

df['dependents'].value_counts(normalize=True, dropna=False)

df['dependents'].isna().sum()

df['dependents'].skew()

df['dependents'].mean(), df['dependents'].median(), df['dependents'].mode()

gp = df.groupby(df['occupation'])
gp['dependents'].mean()
# Avg no. of depedents for each occupation type

gp = df.groupby(df['churn'])
gp['dependents'].mean()
# Avg no. of depedents for each churn type

gp = df.groupby(df['gender'])
gp['dependents'].mean().round()

# wherever the dependents is NAN & gender is male, replace it with 1
# wherever the dependents is NAN & gender is female, replace it with 0

# Implement the strategy no. 2
# wherever the dependents is NAN & gender is male, replace it with 1
# wherever the dependents is NAN & gender is female, replace it with 0

def impute_nan_dependents(cols):
    gender = cols[0]
    dep = cols[1]
    
    if pd.isnull(dep):
        if gender == "Male": 
            return 1  # replace the NaN with 1, when gender is Male
        elif gender == "Female":
            return 0  # replace the NaN with 0, when gender is Female
    else:
        return dep # return the same value of dependents if it is not null

df[['gender', 'dependents']].head()

df['dependents'] = df[['gender', 'dependents']].apply(impute_nan_dependents, axis=1)

df[['gender', 'dependents']].head()

df.isna().sum()

df.tail()

# dropping the customer_id col
df.drop("customer_id", axis=1, inplace=True)

df.head()

df[df.churn.isnull()]  # fetch the row where Churn is missing

df[df.churn.isnull()].index[0]

# Drop those rows where churn is null
df.drop(df[df.churn.isnull()].index[0], axis=0, inplace=True)

df.tail()

df.isna().sum()

df.to_pickle("churn_prediction_v2.pkl")

df.drop(['gender', 'dependents'], axis=1, inplace=True)

df.head()

df = pd.read_pickle("churn_prediction_v2.pkl")
df.head()

# Final missing vlaue treatment

df.dropna(axis=1, inplace=True)  # drop those col which have missing values

df.isna().sum()

df.to_pickle("churn_prediction_v3.pkl")

import pandas as pd
df = pd.read_pickle("churn_prediction_v3.pkl")
df.head()
